add armortype to item generation

+++ InfiniteCreativity/Models/Player.cs
@@ -7,7 +7,7 @@ namespace InfiniteCreativity.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Password { get; set; }
-        public int Purse { get; set; }
+        public double Purse { get; set; }
         public IEnumerable<Item>? Inventory { get; set; }
         public ICollection<Character> Characters { get; set; }
     }
